DevTools Researcher/requirements.txt
firecrawl-py
pydentic
langchain
langchain-openai
langgraph
dotenv